main/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
sqlparse==0.4.2
celery==4.4.2
python-dotenv==0.20.0
djangorestframework==3.13.1
psycopg2-binary==2.9.3
paho-mqtt==1.6.1
redis==4.3.4
